* Converting exceptions to MessagingException on AMQP service client. * Adding EPH with nonexistent entity unit test. * Moving negative tests to dedicated test classes.
