+++ test/Microsoft.Azure.EventHubs.UnitTests/EventHubClientTests.cs
@@ -7,15 +7,15 @@
     using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
-    using Newtonsoft.Json;
     using Xunit;
     using Xunit.Abstractions;
 
     public class EventHubClientTests
     {
         ITestOutputHelper output;
-        string connectionString;
-        string[] PartitionIds;
+        protected string EventHubConnectionString;
+        protected string[] PartitionIds;
+        protected EventHubClient EventHubClient;
 
         public EventHubClientTests(ITestOutputHelper output)
         {
@@ -29,7 +29,7 @@
             // Update operation timeout on ConnectionBuilderString.
             var cbs = new EventHubsConnectionStringBuilder(connectionString);
             cbs.OperationTimeout = TimeSpan.FromSeconds(15);
-            this.connectionString = cbs.ToString();
+            this.EventHubConnectionString = cbs.ToString();
 
             // Create default EH client.
             this.EventHubClient = EventHubClient.CreateFromConnectionString(connectionString);
@@ -40,14 +40,12 @@
             Log($"EventHub has {PartitionIds.Length} partitions");
         }
 
-        EventHubClient EventHubClient { get; }
-
         [Fact]
         void ConnectionStringBuilderTest()
         {
-            Log($"Original connection string: {this.connectionString}");
+            Log($"Original connection string: {this.EventHubConnectionString}");
 
-            var csb = new EventHubsConnectionStringBuilder(this.connectionString);
+            var csb = new EventHubsConnectionStringBuilder(this.EventHubConnectionString);
 
             // Try update settings and rebuild the connection string.
             csb.Endpoint = new Uri("sb://newendpoint");
@@ -116,20 +114,6 @@
                 await pReceiver.ReceiveAsync(1);
                 throw new InvalidOperationException("Receive should have failed");
             });
-        }
-
-        [Fact]
-        Task CreateClientWithoutEntityPathShouldFail()
-        {
-            // Remove entity path from connection string.
-            var csb = new EventHubsConnectionStringBuilder(this.connectionString);
-            csb.EntityPath = null;
-
-            return Assert.ThrowsAsync<ArgumentException>(() =>
-             {
-                 EventHubClient.CreateFromConnectionString(csb.ToString());
-                 throw new Exception("Entity path wasn't provided in the connection string and this new call was supposed to fail");
-             });
         }
 
         [Fact]
@@ -607,65 +591,6 @@
         }
 
         [Fact]
-        async Task MessageSizeExceededException()
-        {
-            try
-            { 
-                Log("Sending large event via EventHubClient.SendAsync(EventData)");
-                var eventData = new EventData(new byte[300000]);
-                await this.EventHubClient.SendAsync(eventData);
-                throw new InvalidOperationException("Send should have failed with " +
-                    typeof(MessageSizeExceededException).Name);
-            }
-            catch (MessageSizeExceededException)
-            {
-                Log("Caught MessageSizeExceededException as expected");
-            }
-        }
-
-        [Fact]
-        async Task SendReceiveNonexistentEntity()
-        {
-            // Rebuild connection string with a nonexistent entity.
-            var csb = new EventHubsConnectionStringBuilder(this.connectionString);
-            csb.EntityPath = Guid.NewGuid().ToString();
-            var ehClient = EventHubClient.CreateFromConnectionString(csb.ToString());
-
-            // Try sending.
-            PartitionSender sender = null;
-            await Assert.ThrowsAsync<MessagingEntityNotFoundException>(async () =>
-            {
-                Log("Sending an event to nonexistent entity.");
-                sender = ehClient.CreatePartitionSender("0");
-                await sender.SendAsync(new EventData(Encoding.UTF8.GetBytes("this send should fail.")));
-                throw new InvalidOperationException("Send should have failed");
-            });
-            await sender.CloseAsync();
-
-            // Try receiving.
-            PartitionReceiver receiver = null;
-            await Assert.ThrowsAsync<MessagingEntityNotFoundException>(async () =>
-            {
-                Log("Receiving from nonexistent entity.");
-                receiver = ehClient.CreateReceiver(PartitionReceiver.DefaultConsumerGroupName, "0", PartitionReceiver.StartOfStream);
-                await receiver.ReceiveAsync(1);
-                throw new InvalidOperationException("Receive should have failed");
-            });
-            await receiver.CloseAsync();
-
-            // Try receiving on an nonexistent consumer group.
-            ehClient = EventHubClient.CreateFromConnectionString(this.connectionString);
-            await Assert.ThrowsAsync<MessagingEntityNotFoundException>(async () =>
-            {
-                Log("Receiving from nonexistent consumer group.");
-                receiver = ehClient.CreateReceiver(Guid.NewGuid().ToString(), "0", PartitionReceiver.StartOfStream);
-                await receiver.ReceiveAsync(1);
-                throw new InvalidOperationException("Receive should have failed");
-            });
-            await receiver.CloseAsync();
-        }
-
-        [Fact]
         async Task PartitionKeyValidation()
         {
             int NumberOfMessagesToSend = 100;
@@ -799,7 +724,7 @@
             return messages;
         }
 
-        void Log(string message)
+        protected void Log(string message)
         {
             var log = string.Format("{0} {1}", DateTime.Now.TimeOfDay, message);
             output.WriteLine(log);
